converted GT-11155 to enyo 2.6
