+++ test/moonstone/Popup/GT-11155-DirectPopupHide/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "GT-11155-DirectPopupHide.js"
+}
